refactor DuplicateCreateTable into schemadiff
